Code cleaning : Remove unnecessary casts

+++ plugins/sonar-dbcleaner-plugin/src/main/java/org/sonar/plugins/dbcleaner/ProjectPurgePostJob.java
@@ -35,6 +35,6 @@ public class ProjectPurgePostJob implements PostJob {
   }
 
   public void executeOn(final Project project, SensorContext context) {
-    purgeTask.purge((long) project.getId());
+    purgeTask.purge(project.getId());
   }
 }
