+++ plugins/sonar-squid-java-plugin/src/main/java/org/sonar/plugins/squid/bridges/DesignBridge.java
@@ -70,9 +70,9 @@ public class DesignBridge extends Bridge {
       LOG.debug("{} feedback edges", feedbackEdges.size());
       int tangles = cyclesAndFESSolver.getWeightOfFeedbackEdgeSet();
 
-      savePositiveMeasure(sonarProject, CoreMetrics.PACKAGE_CYCLES, (double) cyclesAndFESSolver.getCycles().size());
-      savePositiveMeasure(sonarProject, CoreMetrics.PACKAGE_FEEDBACK_EDGES, (double) feedbackEdges.size());
-      savePositiveMeasure(sonarProject, CoreMetrics.PACKAGE_TANGLES, (double) tangles);
+      savePositiveMeasure(sonarProject, CoreMetrics.PACKAGE_CYCLES, cyclesAndFESSolver.getCycles().size());
+      savePositiveMeasure(sonarProject, CoreMetrics.PACKAGE_FEEDBACK_EDGES, feedbackEdges.size());
+      savePositiveMeasure(sonarProject, CoreMetrics.PACKAGE_TANGLES, tangles);
       savePositiveMeasure(sonarProject, CoreMetrics.PACKAGE_EDGES_WEIGHT, getEdgesWeight(squidPackages));
 
       String dsmJson = serializeDsm(squid, squidPackages, feedbackEdges);
@@ -103,9 +103,9 @@ public class DesignBridge extends Bridge {
       Set<Edge> feedbackEdges = solver.getEdges();
       int tangles = solver.getWeightOfFeedbackEdgeSet();
 
-      savePositiveMeasure(sonarPackage, CoreMetrics.FILE_CYCLES, (double) cycles.size());
-      savePositiveMeasure(sonarPackage, CoreMetrics.FILE_FEEDBACK_EDGES, (double) feedbackEdges.size());
-      savePositiveMeasure(sonarPackage, CoreMetrics.FILE_TANGLES, (double) tangles);
+      savePositiveMeasure(sonarPackage, CoreMetrics.FILE_CYCLES, cycles.size());
+      savePositiveMeasure(sonarPackage, CoreMetrics.FILE_FEEDBACK_EDGES, feedbackEdges.size());
+      savePositiveMeasure(sonarPackage, CoreMetrics.FILE_TANGLES, tangles);
       savePositiveMeasure(sonarPackage, CoreMetrics.FILE_EDGES_WEIGHT, getEdgesWeight(squidFiles));
 
       String dsmJson = serializeDsm(squid, squidFiles, feedbackEdges);
